Use uintptr_t to better convey type

+++ src/OpenLoco/src/Map/TileManager.cpp
@@ -947,7 +947,7 @@ namespace OpenLoco::World::TileManager
     }
 
     // 0x00462926
-    static bool canConstructWithClearAt(const World::Pos2& pos, uint8_t baseZ, uint8_t clearZ, const QuarterTile& qt, uint32_t clearFunctionLegacy, uint8_t flags)
+    static bool canConstructWithClearAt(const World::Pos2& pos, uint8_t baseZ, uint8_t clearZ, const QuarterTile& qt, uintptr_t clearFunctionLegacy, uint8_t flags)
     {
         auto functionWrapper = [clearFunctionLegacy](const TileElement& el) -> ClearResult {
             registers regs{};
@@ -984,13 +984,13 @@ namespace OpenLoco::World::TileManager
     }
 
     // 0x00462908
-    bool sub_462908(const World::Pos2& pos, uint8_t baseZ, uint8_t clearZ, const QuarterTile& qt, uint32_t clearFunctionLegacy)
+    bool sub_462908(const World::Pos2& pos, uint8_t baseZ, uint8_t clearZ, const QuarterTile& qt, uintptr_t clearFunctionLegacy)
     {
         return canConstructWithClearAt(pos, baseZ, clearZ, qt, clearFunctionLegacy, (1 << 7) | (1 << 0));
     }
 
     // 0x00462917
-    bool sub_462917(const World::Pos2& pos, uint8_t baseZ, uint8_t clearZ, const QuarterTile& qt, uint32_t clearFunctionLegacy)
+    bool sub_462917(const World::Pos2& pos, uint8_t baseZ, uint8_t clearZ, const QuarterTile& qt, uintptr_t clearFunctionLegacy)
     {
         return canConstructWithClearAt(pos, baseZ, clearZ, qt, clearFunctionLegacy, (1 << 0));
     }
